(PDK-944) Enforce WordArray brackets style
The previous default of `percent` was enforcing a obscure part of the
ruby language that was hostile to new and experienced users alike.

Changing the enforced style to `brackets` avoids folks having to
reformat larger arrays when adding non-word elements, and avoids
the necessity to know too much ruby.

Thanks to Johnson Earls, and others on the #puppet-dev slack for
providing critical feedback, and starting this change.