(MODULES-7184) replace parallel_spec with spec
parallel_spec and bundler 1.16.2 are not playing nicely. The new bundler version was released yesterday and now checks that use parallel_spec are failing. Not sure if this is a permanent solution or not but it is definitely the easiest solution at this point and it should get things green again. Another equally easy option would be to remove the bundler update from the before_install array, but that seems like it would have much more of an impact.