Change the provider example to use the context
The underscore prefix in the `get` method's signature to signify that
the context is not used in the method can be confusing to newcomers.

Instead use the context in the method to print an explanatory message.

This solves the original problem, and adds a hint for what is going on.