Fail on broken spec_helper_local
Since we already check whether the file is there, a `LoadError` will only occur when there is a error in the `spec_helper_local`. A warning is not appropriate in this case, as it will usually only mean that there is another (now obscured) error later in the test run.