(PDK-1210) setting inherited const_defined lookup to false
If there is a top level module with same name as a type which gets loaded before the type gets defined.

This will trigger a false positive on the `last_module.const_defined?` check causing an `uninitialized constant` error.

This change will no longer do a recursive check and only care about the local namespace this way avoiding the false positive.