(MODULES-8854) Set default changelog_version_tag_pattern to empty
The introduction of this config parameter has broken the changelog
links generated by our changelog rake task. This is because our
modules are tagged using the pure semver i.e. '1.2.3' not 'v1.2.3'.

This change resets this back to the original functionality to ensure
changelogs do not contain broken links. Instead, consumers of the
template should define this parameter in their .sync.yml should they
wish to use it, as is the case with all other changelog related
parameters.