(maint) VSCode Recommended Extensions file
This commit adds a VSCode recommended extensions file to the project.
When a user opens up a module managed by the PDK with VSCode, they will
be presented with a popup notification recommending these extensions.
The user can choose to ignore the recommendations or install all or just
one of the extensions.