(PDK-1436) Add circleCI and Azure Pipelines files
This commit adds default YAML files for CircleCI
and Azure Pipelines.  These do not currently have
any way to be customized via .sync.yml.

The purpose of these examples is to provide users
a good default that will work for most and if not
the user can unmanage the file and customize as
needed.  As users make use of these we may
consider adding the ability to make some
customizations via .sync.yml but hopefullly
not as much craziness as travis and gitlab ci
have.