(MODULES-9749) Update changelog instructions
Prior to this commit if the gem for github_changelog_generator could
not be found the error message raised by the changelog rake task
advised the reader to add the changelog generator to the .sync.yaml
in the development gem group.

This commit updates that advice to clarify that the gem may not be
installed intentionally and, if not, to ensure it exists in the gem
file by using .sync.yaml to add it to the release_prep gem group
and running a pdk update.