(PDK-1500) Update Gemfile template for Litmus
Previously if a module used Litmus for acceptance testing it was possible that
Bundler would choose Development metagems that did not include Litmus which
would then cause CI, or local testing failures.  This commit:

* Adds a `use_litmus` setting to the Gemfile
* When the use_litmus setting is set to true, the Gemfile template is modified
  so that the development gems will be select a version that contains the
  litmus gem

Note that this munging is used because it is not possible to add the complex
logic into the config_defaults file. Nor is it possible to configure this using
knockout prefix in a module's sync.yml.