Add starter Forge publish GitHub action
There doesn't seem to be a clean way to enable this action only if the
FORGE_API_TOKEN is set, so users will have to remove the conditional
skipping the job to enable it for now. The `secrets` and the `env`
contexts appear to only be available to individual steps, not the job as
a whole. boo.