(maint) Fix acceptance test setup when not common owner
Prior to this commit, when the user submitting the PR was not the common
owner, the acceptance test setup would fail. The failure was because
there was a missing matrix due to an if statement. This commit generates
an empty matrix and skips acceptance tests if the matrix is empty.