Use ruby/setup-ruby Github action
The ruby/setup-ruby action is actually maintained by the Ruby people
themselves. It has built in caching with the proper behavior. Not only
is it easier to read (less syncing going on), it's also likely to be
maintained properly. For example, to get proper caching of gems, bundle
lock should run to get a Gemfile.lock which gives you a correct cache.
This action does it for you.