(MAINT) Ensure system PMG gem installed when PUPPET_GEM_VERSION nil
Prior to this change, if PUPPET_GEM_VERSION was not set, the system
Puppet Module Gems gem would not be installed, which would not have
installed Litmus. This was causing issues on CI pipelines as there
is a dependency on a helper script in Litmus during set up,
regardless as to whether PUPPET_GEM_VERSION is set or not.

This was introduced in the changes within #385 and #386.

With this change, we now ensure that the system PMG gem is installed
if PUPPET_GEM_VERSION is undefined.