Create changelog if needed
When migrating to gh-changelog-generator, you move your existing
changelog to HISTORY.md and then the generator appends that to the end
of the new CHANGELOG.md file.

However, when this workflow runs, it uses `git diff` to identify changes
in the CHANGELOG.md file, and creation of a new file are not flagged.
The simplest way to make that check work is to create an empty file to
compare against. This should only take effect on the very first release
when HISTORY.md exists, but the new CHANGELOG.md does not.