(CONT-19) Removal of puppet-module-gems
With the recent updates to bundler the need for a specific set of gems for each version of ruby that we support has been removed.
As such the logic behind our gems is now much simpler and can easily fit within the default templates, thus removing our need to have a separate module to manage them, with the templates already managing all other similar default requirements for our modules.