(CAT-434) Update README.md
Previous to this commit, our PDK-template documentation would be fairly
intimadating and confusing, often jumpin from basic introductions to
advanced concepts, while expecting the reader to have an understanding
of concepts they may not be familiar with.

This commit aims to rework the README to ensure that its more readable
and user-friendly.