(CAT-1868) - Remove .rspec.erb
This commit removes the .rspec.erb file from the pdk-templates, and thus
the .rspec file from being managed by PDK. This was made possible due to
the changes found in https://github.com/puppetlabs/puppetlabs_spec_helper/pull/452.

This has minimal impact on users as the file was not configurable
through .sync.yml before (other than to set as unmanaged), existing
modules with the .rspec will not be affected and new modules created will
simply not include an .rspec file.