Only pin to the puppetcore versions.
Since the puppet_version and facter_version are not always going to be version contrains (they can also be a git URL), then I'm removing this flexible check and pinning only to the latest puppetcore gems.  If flexible puppet and facter gemsources are need via https or git, then remove the PUPPET_FORGE_TOKEN and use PUPPET_GEM_VERSION and/or FACTER_GEM_VERSION instead.

Signed-off-by: Gavin Didrichsen <gavin.didrichsen@gmail.com>